requirements.txt
# FastAPI dan dependencies utama
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP dan networking
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis untuk caching
redis==5.0.1
aioredis==2.0.1

# Authentication dan security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0
phonenumbers==8.13.26

# Monitoring dan logging
prometheus-client==0.19.0
structlog==23.2.0

# Task queue
celery==5.3.4
kombu==5.3.4

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8
